// mobile/api/gen/src/models/WeightedLatLngDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Offlinery
 * API of Offlinery
 *
 * The version of the OpenAPI document: 1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface WeightedLatLngDTO
 */
export interface WeightedLatLngDTO {
    /**
     *
     * @type {number}
     * @memberof WeightedLatLngDTO
     */
    latitude: number;
    /**
     *
     * @type {number}
     * @memberof WeightedLatLngDTO
     */
    longitude: number;
    /**
     *
     * @type {number}
     * @memberof WeightedLatLngDTO
     */
    weight?: number;
}

/**
 * Check if a given object implements the WeightedLatLngDTO interface.
 */
export function instanceOfWeightedLatLngDTO(
    value: object,
): value is WeightedLatLngDTO {
    if (!("latitude" in value) || value["latitude"] === undefined) return false;
    if (!("longitude" in value) || value["longitude"] === undefined)
        return false;
    return true;
}

export function WeightedLatLngDTOFromJSON(json: any): WeightedLatLngDTO {
    return WeightedLatLngDTOFromJSONTyped(json, false);
}

export function WeightedLatLngDTOFromJSONTyped(
    json: any,
    ignoreDiscriminator: boolean,
): WeightedLatLngDTO {
    if (json == null) {
        return json;
    }
    return {
        latitude: json["latitude"],
        longitude: json["longitude"],
        weight: json["weight"] == null ? undefined : json["weight"],
    };
}

export function WeightedLatLngDTOToJSON(value?: WeightedLatLngDTO | null): any {
    if (value == null) {
        return value;
    }
    return {
        latitude: value["latitude"],
        longitude: value["longitude"],
        weight: value["weight"],
    };
}